+++ src/unpoly/link.js
@@ -1203,7 +1203,7 @@ up.link = (function() {
     In particular:
 
     - [Animations](/up.motion) have concluded and [transitioned](https://unpoly.com/a-up-transition) elements were removed from the DOM tree.
-    - A [cached response](#options.cache) was [revalidated with the server](/caching#revalidation).
+    - A [cached response](#up-cache) was [revalidated with the server](/caching#revalidation).
       If the server has responded with new content, this content has also been rendered.
 
     | Expression | Value                                                                  |